remove tickImages prop
